requirements.txt
ansicon==1.89.0
arrow==1.2.3
asgiref==3.6.0
attrs==22.2.0
blessed==1.20.0
certifi==2022.12.7
charset-normalizer==3.1.0
Django==4.1.7
django-picklefield==3.1
django-q==1.3.9
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.26.0
flake8==6.0.0
idna==3.4
inflection==0.5.1
jinxed==1.2.0
jsonschema==4.17.3
mccabe==0.7.0
pycodestyle==2.10.0
pyflakes==3.0.1
PyJWT==2.6.0
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.7.1
PyYAML==6.0
redis==3.5.3
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.14
wcwidth==0.2.6
